Reorganizing transaction traits
