--- crates/relayer/relayer-components/src/transaction/traits/nonce/error.rs
@@ -1,16 +1,0 @@
-use cgp_core::CanRaiseError;
-
-use crate::transaction::traits::types::HasNonceType;
-
-pub struct NonceMistmatchError<Nonce> {
-    pub expected_nonce: Nonce,
-    pub given_nonce: Nonce,
-}
-
-pub trait HasNonceMismatchError:
-    HasNonceType + CanRaiseError<NonceMistmatchError<Self::Nonce>>
-{
-    fn try_extract_nonce_mismatch_error(
-        e: &Self::Error,
-    ) -> Option<NonceMistmatchError<Self::Nonce>>;
-}
