+++ crates/relayer/relayer-components/src/transaction/traits/nonce/mod.rs
@@ -1,3 +1,4 @@
-pub mod error;
+pub mod allocate_nonce;
 pub mod guard;
 pub mod mutex;
+pub mod query_nonce;
